# openapi/content-studio-openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/contents":{"post":{"tags":["content-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateContentDraftPayloadRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateContentDraftPayloadResponse"}}}}}}},"/contents/{contentId}/submit-review":{"post":{"tags":["review-controller"],"operationId":"submitReview","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/contents/{contentId}/approve":{"post":{"tags":["review-controller"],"operationId":"approve","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApproveContentReviewPayloadRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/advanced/contents/{contentId}/release-readiness/manual-confirm":{"post":{"tags":["advanced-release-readiness-controller"],"operationId":"confirmManual","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/advanced/contents/{contentId}/release-readiness/copyright-pass":{"post":{"tags":["advanced-release-readiness-controller"],"operationId":"passCopyright","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/advanced/contents/gated":{"post":{"tags":["advanced-release-readiness-controller"],"operationId":"createGated","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGatedContentDraftPayloadRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateGatedContentDraftPayloadResponse"}}}}}}},"/media-processing/{contentId}":{"get":{"tags":["query-controller"],"operationId":"getMediaProcessing","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetMediaProcessingStatusPayloadResponse"}}}}}}},"/contents/{contentId}":{"get":{"tags":["query-controller"],"operationId":"getContent","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetContentDetailPayloadResponse"}}}}}}}},"components":{"schemas":{"CreateContentDraftPayloadRequest":{"type":"object","properties":{"title":{"type":"string"},"body":{"type":"string"},"mediaSourceKey":{"type":"string"}},"required":["body","mediaSourceKey","title"]},"CreateContentDraftPayloadResponse":{"type":"object","properties":{"contentId":{"type":"string","format":"uuid"}},"required":["contentId"]},"ApproveContentReviewPayloadRequest":{"type":"object","properties":{"reviewerId":{"type":"string","format":"uuid"}},"required":["reviewerId"]},"CreateGatedContentDraftPayloadRequest":{"type":"object","properties":{"title":{"type":"string"},"body":{"type":"string"},"mediaSourceKey":{"type":"string"},"releaseWindowOpensAt":{"type":"string","format":"date-time"},"releaseWindowClosesAt":{"type":"string","format":"date-time"}},"required":["body","mediaSourceKey","releaseWindowClosesAt","releaseWindowOpensAt","title"]},"CreateGatedContentDraftPayloadResponse":{"type":"object","properties":{"contentId":{"type":"string","format":"uuid"}},"required":["contentId"]},"GetMediaProcessingStatusPayloadResponse":{"type":"object","properties":{"contentId":{"type":"string","format":"uuid"},"task":{"$ref":"#/components/schemas/GetMediaProcessingStatusPayloadResponseTask"}},"required":["contentId"]},"GetMediaProcessingStatusPayloadResponseTask":{"type":"object","properties":{"taskId":{"type":"string","format":"uuid"},"externalTaskId":{"type":"string"},"processingStatus":{"type":"string"}},"required":["processingStatus","taskId"]},"GetContentDetailPayloadResponse":{"type":"object","properties":{"contentId":{"type":"string","format":"uuid"},"title":{"type":"string"},"body":{"type":"string"},"mediaSourceKey":{"type":"string"},"reviewStatus":{"type":"string"},"contentStatus":{"type":"string"},"reviewerId":{"type":"string","format":"uuid"},"reviewedAt":{"type":"string","format":"date-time"},"publishedAt":{"type":"string","format":"date-time"}},"required":["body","contentId","contentStatus","mediaSourceKey","reviewStatus","title"]}}}}